+++ src/SMAPI.Toolkit/ModToolkit.cs
@@ -22,7 +22,7 @@ namespace StardewModdingAPI.Toolkit
         private readonly string UserAgent;
 
         /// <summary>Maps vendor keys (like <c>Nexus</c>) to their mod URL template (where <c>{0}</c> is the mod ID). This doesn't affect update checks, which defer to the remote web API.</summary>
-        private readonly IDictionary<string, string> VendorModUrls = new Dictionary<string, string>(StringComparer.InvariantCultureIgnoreCase)
+        private readonly IDictionary<string, string> VendorModUrls = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
         {
             ["Chucklefish"] = "https://community.playstarbound.com/resources/{0}",
             ["GitHub"] = "https://github.com/{0}/releases",