+++ src/SMAPI/Framework/Content/AssetDataForMap.cs
@@ -67,7 +67,7 @@ namespace StardewModdingAPI.Framework.Content
                 {
                     // change ID if needed so new tilesheets are added after vanilla ones (to avoid errors in hardcoded game logic)
                     string id = sourceSheet.Id;
-                    if (!id.StartsWith("z_", StringComparison.InvariantCultureIgnoreCase))
+                    if (!id.StartsWith("z_", StringComparison.OrdinalIgnoreCase))
                         id = $"z_{id}";
 
                     // change ID if it conflicts with an existing tilesheet